requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
# langchain-ollama>=0.3.0
# langchain-xai>=0.1.0
langgraph>=0.2.0
langchain-community>=0.3.0
tavily-python>=0.5.0
# exa-py>=1.0.0
# google-search-results>=2.4.0
python-dotenv>=1.0.0
pydantic>=2.7.0
streamlit>=1.25.0
